+++ js/utils/check-max-string-length.js
@@ -1,0 +1,5 @@
+const checkMaxStringLength = (string, maxLength) => {
+  return string.length <= maxLength;
+};
+
+export { checkMaxStringLength };
